+++ src/main/java/org/opentripplanner/standalone/config/ItineraryFiltersMapper.java
@@ -15,6 +15,7 @@ public class ItineraryFiltersMapper {
         c.asBoolean("debug", dft.debug),
         c.asDouble("groupSimilarityKeepOne", dft.groupSimilarityKeepOne),
         c.asDouble("groupSimilarityKeepThree", dft.groupSimilarityKeepThree),
+        c.asDouble("groupedOtherThanSameLegsMaxCostMultiplier", dft.groupedOtherThanSameLegsMaxCostMultiplier),
         c.asDouble("minSafeTransferTimeFactor", dft.minSafeTransferTimeFactor),
         c.asLinearFunction("transitGeneralizedCostLimit", dft.transitGeneralizedCostLimit),
         c.asLinearFunction("nonTransitGeneralizedCostLimit", dft.nonTransitGeneralizedCostLimit),