+++ src/main/java/org/opentripplanner/transit/raptor/rangeraptor/transit/TripScheduleAlightSearch.java
@@ -3,6 +3,7 @@ package org.opentripplanner.transit.raptor.rangeraptor.transit;
 import javax.annotation.Nullable;
 import org.opentripplanner.model.base.ToStringBuilder;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTimeTable;
+import org.opentripplanner.transit.raptor.api.transit.RaptorTransferConstraint;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripSchedule;
 import org.opentripplanner.transit.raptor.api.transit.RaptorTripScheduleBoardOrAlightEvent;
 
@@ -45,24 +46,28 @@ public final class TripScheduleAlightSearch<T extends RaptorTripSchedule>
     /* TripScheduleBoardOrAlightEvent implementation using fly-weight pattern */
 
     @Override
-    public final T getTrip() {
+    public T getTrip() {
         return candidateTrip;
     }
 
     @Override
-    public final int getTripIndex() {
+    public int getTripIndex() {
         return candidateTripIndex;
     }
 
     @Override
-    public final int getTime() {
+    public int getTime() {
         return candidateTrip.arrival(stopPositionInPattern);
     }
 
     @Override
-    public final int getStopPositionInPattern() {
+    public int getStopPositionInPattern() {
         return stopPositionInPattern;
     }
+
+    @Nullable
+    @Override
+    public RaptorTransferConstraint getTransferConstraint() { return null; }
 
 
     /* TripScheduleSearch implementation */
@@ -76,7 +81,7 @@ public final class TripScheduleAlightSearch<T extends RaptorTripSchedule>
      * @param tripIndexLowerBound   Upper bound for trip index to search for (exclusive).
      */
     @Override
-    public final RaptorTripScheduleBoardOrAlightEvent<T> search(
+    public RaptorTripScheduleBoardOrAlightEvent<T> search(
             int latestAlightTime,
             int stopPositionInPattern,
             int tripIndexLowerBound