+++ src/test/java/org/opentripplanner/transit/raptor/rangeraptor/path/PathMapperTest.java
@@ -2,33 +2,39 @@ package org.opentripplanner.transit.raptor.rangeraptor.path;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.BASIC_PATH_AS_DETAILED_STRING;
+import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.COST_CALCULATOR;
 import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.SLACK_PROVIDER;
 import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.basicTripByForwardSearch;
 import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.basicTripByReverseSearch;
 import static org.opentripplanner.transit.raptor._data.stoparrival.BasicPathTestCase.lifeCycle;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAForwardSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAForwardSearchAsText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAReverseSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAReverseSearchAsText;
+import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAWithOpeningHoursForwardSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAWithOpeningHoursForwardSearchAsText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAWithOpeningHoursReverseSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAWithOpeningHoursReverseSearchAsText;
+import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseAWithOpeningHoursText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBForwardSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBForwardSearchAsText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBReverseSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBReverseSearchAsText;
+import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBWithOpeningHoursForwardSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBWithOpeningHoursForwardSearchAsText;
 import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBWithOpeningHoursReverseSearch;
-import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBWithOpeningHoursReverseSearchAsText;
+import static org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase.flexCaseBWithOpeningHoursText;
 
 import org.junit.jupiter.api.Test;
+import org.opentripplanner.routing.algorithm.raptor.transit.cost.DefaultCostCalculator;
 import org.opentripplanner.transit.raptor._data.RaptorTestConstants;
+import org.opentripplanner.transit.raptor._data.stoparrival.FlexAccessAndEgressPathTestCase;
 import org.opentripplanner.transit.raptor._data.transit.TestTripSchedule;
 import org.opentripplanner.transit.raptor.api.path.Path;
+import org.opentripplanner.transit.raptor.api.transit.RaptorSlackProvider;
 
 public class PathMapperTest implements RaptorTestConstants {
+
+    private static final RaptorSlackProvider FLEX_SLACK_PROVIDER =
+            FlexAccessAndEgressPathTestCase.SLACK_PROVIDER;
+    private static final DefaultCostCalculator FLEX_COST_CALCULATOR =
+            FlexAccessAndEgressPathTestCase.COST_CALCULATOR;
+
 
     /* BASIC CASES */
 
@@ -36,7 +42,14 @@ public class PathMapperTest implements RaptorTestConstants {
     public void mapToPathBasicForwardSearch() {
         // Given:
         var destArrival = basicTripByForwardSearch();
-        var mapper = new ForwardPathMapper<TestTripSchedule>(SLACK_PROVIDER, lifeCycle());
+        var mapper = new ForwardPathMapper<TestTripSchedule>(
+                null,
+                SLACK_PROVIDER,
+                COST_CALCULATOR,
+                this::stopIndexToName,
+                lifeCycle(),
+                false
+        );
 
         //When:
         Path<TestTripSchedule> path = mapper.mapToPath(destArrival);
@@ -49,7 +62,14 @@ public class PathMapperTest implements RaptorTestConstants {
     public void mapToPathBasicReverseSearch() {
         // Given:
         var destArrival = basicTripByReverseSearch();
-        var mapper = new ReversePathMapper<TestTripSchedule>(SLACK_PROVIDER, lifeCycle());
+        var mapper = new ReversePathMapper<TestTripSchedule>(
+                null,
+                SLACK_PROVIDER,
+                COST_CALCULATOR,
+                this::stopIndexToName,
+                lifeCycle(),
+                false
+        );
 
         //When:
         Path<TestTripSchedule> path = mapper.mapToPath(destArrival);
@@ -63,27 +83,27 @@ public class PathMapperTest implements RaptorTestConstants {
 
     @Test
     public void mapToPathForFlexCaseAForwardSearch() {
-        runtTestFlexForward(flexCaseAForwardSearch(), flexCaseAForwardSearchAsText());
+        runTestFlexForward(flexCaseAForwardSearch(), flexCaseAText());
     }
 
     @Test
     public void mapToPathForFlexCaseAWOpeningHoursForwardSearch() {
-        runtTestFlexForward(
+        runTestFlexForward(
                 flexCaseAWithOpeningHoursForwardSearch(),
-                flexCaseAWithOpeningHoursForwardSearchAsText()
+                flexCaseAWithOpeningHoursText()
         );
     }
 
     @Test
     public void mapToPathForFlexCaseBForwardSearch() {
-        runtTestFlexForward(flexCaseBForwardSearch(), flexCaseBForwardSearchAsText());
+        runTestFlexForward(flexCaseBForwardSearch(), flexCaseBText());
     }
 
     @Test
     public void mapToPathForFlexCaseBWOpeningHoursForwardSearch() {
-        runtTestFlexForward(
+        runTestFlexForward(
                 flexCaseBWithOpeningHoursForwardSearch(),
-                flexCaseBWithOpeningHoursForwardSearchAsText()
+                flexCaseBWithOpeningHoursText()
         );
     }
 
@@ -91,55 +111,75 @@ public class PathMapperTest implements RaptorTestConstants {
 
     @Test
     public void mapToPathForFlexCaseAReverseSearch() {
-        runtTestFlexReverse(flexCaseAReverseSearch(), flexCaseAReverseSearchAsText());
+        runTestFlexReverse(flexCaseAReverseSearch(), flexCaseAText());
     }
 
     @Test
     public void mapToPathForFlexCaseAWOpeningHoursReverseSearch() {
-        runtTestFlexReverse(
+        runTestFlexReverse(
                 flexCaseAWithOpeningHoursReverseSearch(),
-                flexCaseAWithOpeningHoursReverseSearchAsText()
+                flexCaseAWithOpeningHoursText()
         );
     }
 
     @Test
     public void mapToPathForFlexCaseBReverseSearch() {
-        runtTestFlexReverse(flexCaseBReverseSearch(), flexCaseBReverseSearchAsText());
+        runTestFlexReverse(flexCaseBReverseSearch(), flexCaseBText());
     }
 
     @Test
     public void mapToPathForFlexCaseBWOpeningHoursReverseSearch() {
-        runtTestFlexReverse(
+        runTestFlexReverse(
                 flexCaseBWithOpeningHoursReverseSearch(),
-                flexCaseBWithOpeningHoursReverseSearchAsText()
+                flexCaseBWithOpeningHoursText()
         );
     }
 
-    private void runtTestFlexForward(
-            DestinationArrival<TestTripSchedule> destArrival,
-            String expected
-    ) {
-        // Given:
-        var mapper = new ForwardPathMapper<TestTripSchedule>(SLACK_PROVIDER, lifeCycle());
-        // When:
-        Path<TestTripSchedule> path = mapper.mapToPath(destArrival);
-        // Then:
-        assertEquals(expected, path.toStringDetailed(this::stopIndexToName));
-    }
 
-    private void runtTestFlexReverse(
-            DestinationArrival<TestTripSchedule> destArrival,
-            String expected
-    ) {
-        // Given:
-        var mapper = new ReversePathMapper<TestTripSchedule>(SLACK_PROVIDER, lifeCycle());
-        // When:
-        Path<TestTripSchedule> path = mapper.mapToPath(destArrival);
-        // Then:
-        assertEquals(expected, path.toStringDetailed(this::stopIndexToName));
-    }
+    /* private helper methods */
 
     private void assertPath(Path<TestTripSchedule> path) {
-        assertEquals(BASIC_PATH_AS_DETAILED_STRING, path.toStringDetailed(this::stopIndexToName));
+        assertEquals(
+                BASIC_PATH_AS_DETAILED_STRING,
+                path.toStringDetailed(this::stopIndexToName)
+        );
+    }
+
+    private void runTestFlexForward(
+            DestinationArrival<TestTripSchedule> destArrival,
+            String expected
+    ) {
+        // Given:
+        var mapper = new ForwardPathMapper<TestTripSchedule>(
+                null,
+                FLEX_SLACK_PROVIDER,
+                FLEX_COST_CALCULATOR,
+                this::stopIndexToName,
+                lifeCycle(),
+                false
+        );
+        // When:
+        Path<TestTripSchedule> path = mapper.mapToPath(destArrival);
+        // Then:
+        assertEquals(expected, path.toStringDetailed(this::stopIndexToName));
+    }
+
+    private void runTestFlexReverse(
+            DestinationArrival<TestTripSchedule> destArrival,
+            String expected
+    ) {
+        // Given:
+        var mapper = new ReversePathMapper<TestTripSchedule>(
+                null,
+                FLEX_SLACK_PROVIDER,
+                FLEX_COST_CALCULATOR,
+                this::stopIndexToName,
+                lifeCycle(),
+                false
+        );
+        // When:
+        Path<TestTripSchedule> path = mapper.mapToPath(destArrival);
+        // Then:
+        assertEquals(expected, path.toStringDetailed(this::stopIndexToName));
     }
 }